// src/app/models/new.ts
export interface New {
    id_new?: number;
    id_new_type?: number;
    id_user?: number;
    id_class?: number;
    class_name?: string;
    last_name?: string;
    name?: string;
    second_last_name?: string;
    title?: string;
    description?: string;
    date?: string;
    image?: string;
}
//class_name: "1-1"
//date: "2019-09-04"
//description: "Porque me dio flojera ir"
//id_class: 9
//id_new: 3
//id_new_type: 2
//id_user: 10
//image: "/**/"
//last_name: "Campos"
//name: "Daniela"
//second_last_name: "Alfaro"

//title: "Se suspenden clases"